+++ js/src/builtin/Generator.js
@@ -2,16 +2,16 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-function StarGeneratorNext(val) {
-    // The IsSuspendedStarGenerator call below is not necessary for
-    // correctness. It's a performance optimization to check for the
-    // common case with a single call. It's also inlined in Baseline.
+function GeneratorNext(val) {
+    // The IsSuspendedGenerator call below is not necessary for correctness.
+    // It's a performance optimization to check for the common case with a
+    // single call. It's also inlined in Baseline.
 
-    if (!IsSuspendedStarGenerator(this)) {
-        if (!IsObject(this) || !IsStarGeneratorObject(this))
-            return callFunction(CallStarGeneratorMethodIfWrapped, this, val, "StarGeneratorNext");
+    if (!IsSuspendedGenerator(this)) {
+        if (!IsObject(this) || !IsGeneratorObject(this))
+            return callFunction(CallGeneratorMethodIfWrapped, this, val, "GeneratorNext");
 
-        if (StarGeneratorObjectIsClosed(this))
+        if (GeneratorObjectIsClosed(this))
             return { value: undefined, done: true };
 
         if (GeneratorIsRunning(this))
@@ -21,18 +21,18 @@ function StarGeneratorNext(val) {
     try {
         return resumeGenerator(this, val, "next");
     } catch (e) {
-        if (!StarGeneratorObjectIsClosed(this))
+        if (!GeneratorObjectIsClosed(this))
             GeneratorSetClosed(this);
         throw e;
     }
 }
 
-function StarGeneratorThrow(val) {
-    if (!IsSuspendedStarGenerator(this)) {
-        if (!IsObject(this) || !IsStarGeneratorObject(this))
-            return callFunction(CallStarGeneratorMethodIfWrapped, this, val, "StarGeneratorThrow");
+function GeneratorThrow(val) {
+    if (!IsSuspendedGenerator(this)) {
+        if (!IsObject(this) || !IsGeneratorObject(this))
+            return callFunction(CallGeneratorMethodIfWrapped, this, val, "GeneratorThrow");
 
-        if (StarGeneratorObjectIsClosed(this))
+        if (GeneratorObjectIsClosed(this))
             throw val;
 
         if (GeneratorIsRunning(this))
@@ -42,18 +42,18 @@ function StarGeneratorThrow(val) {
     try {
         return resumeGenerator(this, val, "throw");
     } catch (e) {
-        if (!StarGeneratorObjectIsClosed(this))
+        if (!GeneratorObjectIsClosed(this))
             GeneratorSetClosed(this);
         throw e;
     }
 }
 
-function StarGeneratorReturn(val) {
-    if (!IsSuspendedStarGenerator(this)) {
-        if (!IsObject(this) || !IsStarGeneratorObject(this))
-            return callFunction(CallStarGeneratorMethodIfWrapped, this, val, "StarGeneratorReturn");
+function GeneratorReturn(val) {
+    if (!IsSuspendedGenerator(this)) {
+        if (!IsObject(this) || !IsGeneratorObject(this))
+            return callFunction(CallGeneratorMethodIfWrapped, this, val, "GeneratorReturn");
 
-        if (StarGeneratorObjectIsClosed(this))
+        if (GeneratorObjectIsClosed(this))
             return { value: val, done: true };
 
         if (GeneratorIsRunning(this))
@@ -62,9 +62,9 @@ function StarGeneratorReturn(val) {
 
     try {
         var rval = { value: val, done: true };
-        return resumeGenerator(this, rval, "close");
+        return resumeGenerator(this, rval, "return");
     } catch (e) {
-        if (!StarGeneratorObjectIsClosed(this))
+        if (!GeneratorObjectIsClosed(this))
             GeneratorSetClosed(this);
         throw e;
     }
@@ -80,6 +80,6 @@ function InterpretGeneratorResume(gen, val, kind) {
        return resumeGenerator(gen, val, "next");
     if (kind === "throw")
        return resumeGenerator(gen, val, "throw");
-    assert(kind === "close", "Invalid resume kind");
-    return resumeGenerator(gen, val, "close");
+    assert(kind === "return", "Invalid resume kind");
+    return resumeGenerator(gen, val, "return");
 }