+++ js/src/vm/SelfHosting.h
@@ -53,6 +53,9 @@ intrinsic_NewArrayIterator(JSContext* cx, unsigned argc, JS::Value* vp);
 bool
 intrinsic_NewStringIterator(JSContext* cx, unsigned argc, JS::Value* vp);
 
+bool
+intrinsic_IsSuspendedGenerator(JSContext* cx, unsigned argc, JS::Value* vp);
+
 } /* namespace js */
 
 #endif /* vm_SelfHosting_h_ */